Refactor: Post 엔티티 내 blind, deleteCheck 자료형 수정 byte -> int 로 수정

+++ src/main/java/com/example/fieldpasserbe/post/entity/Post.java
@@ -76,10 +76,10 @@ public class Post {
     private Integer wishCount;
 
     @Column(name = "BLIND" , columnDefinition = "TINYINT(1) DEFAULT 0", length = 1)
-    private byte blind;
+    private int blind;
 
     @Column(name = "DELETE_CHECK", columnDefinition = "TINYINT(1) DEFAULT 0", length = 1)
-    private byte deleteCheck;
+    private int deleteCheck;
 
     @PrePersist
     public void prePersist() {
